requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
asn1crypto==1.5.1
cbor2==5.5.1
certifi==2023.11.17
certvalidator==0.11.1
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Django==4.2.9
django-appconf==1.0.6
django-cors-headers==4.3.1
django-countries==7.5.1
django-webauthn==0.0.2
djangorestframework==3.14.0
idna==3.6
oscrypto==1.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyOpenSSL==23.3.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
webauthn==1.11.1

scipy~=1.12.0